workspace/20250808_005438_ocr_poc/code/requirements.txt
opencv-python
easyocr
numpy
pillow
logging